feat: Read data from API, cache it in indexedDB
In the second stage of the project, data is fetched from an API server, and it's saved in indexedDB using the local forage library. The rest of the app logic is the same of stage 1 but I've added some comments to the code and formatted it, I've added a manifest to pass PWA audits and icons.