feat: read reviews from API server
In the stage 2 of the project, reviews were saved as a property of the restaurant object.
The stage 3 API server instead provide reviews at a different endpoint.
Reviews are fetched and cached in indexedDB when the main page have finished loading, in order to achieve a faster restaurant page load.
When the user navigate to the restaurant page reviews should be chached, but I've added a check on indexedDB data, in order to fetch them just in case the user is landing on restaurant page on first visit.