feat: Add favorite restaurants toggle
I've added a button on the restaurant details page to toggle favorite status.
The initial status is fetched from the API server and cached on IDB, and when the user toggle the button the is_favorite property of the restaurant is cached on IDB and thensent to the API server.